Improve docs; move `ReportFilter`
