+++ core/lib/instrument/src/lib.rs
@@ -1,3 +1,4 @@
-//! Shared general-purpose instrumentation used by ZKsync node.
+//! General-purpose instrumentation used by ZKsync libraries.
 
 pub mod alloc;
+pub mod filter;
