+++ core/node/api_server/Cargo.toml
@@ -19,6 +19,7 @@ zksync_contracts.workspace = true
 zksync_types.workspace = true
 zksync_dal = { workspace = true, features = ["node_framework"] }
 zksync_health_check = { workspace = true, features = ["node_framework"] }
+zksync_instrument.workspace = true
 zksync_node_fee_model.workspace = true
 zksync_object_store = { workspace = true, features = ["node_framework"] }
 zksync_shared_metrics.workspace = true
