+++ core/node/api_server/src/web3/metrics.rs
@@ -6,6 +6,7 @@ use vise::{
     Buckets, Counter, DurationAsSecs, EncodeLabelSet, EncodeLabelValue, Family, Gauge, Histogram,
     Info, LabeledFamily, Metrics, MetricsFamily, Unit,
 };
+use zksync_instrument::filter::{report_filter, ReportFilter};
 use zksync_types::api;
 use zksync_web3_decl::error::Web3Error;
 
@@ -13,7 +14,7 @@ use super::{
     backend_jsonrpsee::MethodMetadata, ApiTransport, InternalApiConfig, OptionalApiParams,
     TypedFilter,
 };
-use crate::{tx_sender::SubmitTxError, utils::ReportFilter};
+use crate::tx_sender::SubmitTxError;
 
 /// Observed version of RPC parameters. Have a bounded upper-limit size (256 bytes), so that we don't over-allocate.
 #[derive(Debug)]
